requirements.txt
art==6.1
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.0
python-dotenv==1.0.1
tabulate==0.9.0
